python-mysql/py_todolist_be/requirements.txt
pydantic==1.8.2
SQLAlchemy==1.4.23
uvicorn==0.15.0
Flask==1.1.2
fastapi==0.68.1
fastapi_utils==0.2.1
python-dotenv==0.19.0
